+++ core/src/ons/util.js
@@ -564,4 +564,22 @@ util.checkMissingImport = (...elementNames) => {
   });
 }
 
+/**
+ * On a given object, defines a boolean property that reflects an attribute of the same name.
+ */
+util.defineBooleanProperty = (object, propertyName) => {
+  Object.defineProperty(object, propertyName, {
+    get() {
+      return this.hasAttribute(propertyName);
+    },
+    set(value) {
+      if (value) {
+        this.setAttribute(propertyName, '');
+      } else {
+        this.removeAttribute(propertyName);
+      }
+    }
+  });
+}
+
 export default util;